// apps/frontend/src/api/generated/model/getRelationshipsQueryParams.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Instech API
 * REST API for Instech.
 * This API is used to manage tools and their relationships.
 *
 * OpenAPI spec version: 1.0.0
 */
import type { RelationshipKind } from './relationshipKind.ts';

export type GetRelationshipsQueryParams = {
/**
 * Filter by tool numeric ID
 */
toolId?: number;
/**
 * Filter by relationship kind
 */
kind?: RelationshipKind;
/**
 * Pagination cursor
 */
cursor?: number;
/**
 * Page size limit
 */
limit?: number;
};